Cleaned up some redundant casts and deprecated calls.
git-svn-id: https://samskivert.googlecode.com/svn/trunk@2243 6335cc39-0255-0410-8fd6-9bcaacd3b74c

+++ src/java/com/samskivert/swing/ScaledIcon.java
@@ -64,13 +64,13 @@ public class ScaledIcon implements Icon
     // documentation inherited from interface Icon
     public int getIconWidth ()
     {
-        return (int) Math.round(_icon.getIconWidth() * _scale);
+        return Math.round(_icon.getIconWidth() * _scale);
     }
 
     // documentation inherited from interface Icon
     public int getIconHeight ()
     {
-        return (int) Math.round(_icon.getIconHeight() * _scale);
+        return Math.round(_icon.getIconHeight() * _scale);
     }
 
     // documentation inherited from interface Icon
